requirements.txt
amqp==5.0.5
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.1
aspy.refactor-imports==2.2.0
backcall==0.2.0
billiard==3.6.3.0
black==20.8b1
cached-property==1.5.2
celery==5.0.5
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
distlib==0.3.1
Django==3.1.5
django-mysql==3.11.1
django-redis==4.12.1
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
identify==2.1.0
idna==2.10
importlib-metadata==3.7.0
inflection==0.5.1
ipython==7.21.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
nodeenv==1.5.0
packaging==20.9
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==2.10.1
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
seed-isort-config==2.2.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
virtualenv==20.4.2
wcwidth==0.2.5
zipp==3.4.1
